+++ src/win32/TileView.cpp
@@ -529,8 +529,8 @@ LRESULT TileView::OnMapInfo(WPARAM wParam, LPARAM lParam)
   u8 *colors = (u8 *)lParam;
   zoom.setColors(colors);
 
-  int x = (wParam & 0xFFFF)/8;
-  int y = ((wParam >> 16) & 0xFFFF)/8;
+  int x = (int)((wParam & 0xFFFF) / 8);
+  int y = (int)(((wParam >> 16) & 0xFFFF) / 8);
 
   u32 address = 0x6000000 + 0x4000 * charBase;
   int tile = 32 * y + x;
